Create one dedicated file for each type
