+++ common/env-var.go
@@ -1,0 +1,6 @@
+package common
+
+type EnvVar struct {
+	Key   string
+	Value string
+}
