+++ packs/detector-interface.go
