+++ packs/dockerfile-context-base.go
@@ -1,0 +1,8 @@
+package packs
+
+import "github.com/cloud66/starter/common"
+
+type DockerfileContextBase struct {
+	Version  string
+	Packages *common.Lister
+}
