+++ packs/dockerfile-writer-base.go
@@ -9,11 +9,6 @@ import (
 	"github.com/cloud66/starter/common"
 )
 
-type DockerfileContextBase struct {
-	Version  string
-	Packages *common.Lister
-}
-
 type DockerfileWriterBase struct {
 	PackElement
 	TemplateDir     string
