+++ packs/node/dockerfile-context.go
@@ -1,0 +1,7 @@
+package node
+
+import "github.com/cloud66/starter/packs"
+
+type DockerfileContext struct {
+	packs.DockerfileContextBase
+}
