+++ packs/node/dockerfile-writer.go
@@ -2,10 +2,6 @@ package node
 
 import "github.com/cloud66/starter/packs"
 
-type DockerfileContext struct {
-	packs.DockerfileContextBase
-}
-
 type DockerfileWriter struct {
 	packs.DockerfileWriterBase
 }
