+++ packs/node/service-yaml-context.go
@@ -1,0 +1,7 @@
+package node
+
+import "github.com/cloud66/starter/packs"
+
+type ServiceYAMLContext struct {
+	packs.ServiceYAMLContextBase
+}
