+++ packs/pack-base.go
@@ -1,0 +1,7 @@
+package packs
+
+import "github.com/cloud66/starter/common"
+
+type PackBase struct {
+	Messages *common.Lister
+}
