+++ packs/pack-element.go
@@ -1,0 +1,9 @@
+package packs
+
+type PackElement struct {
+	Pack Pack
+}
+
+func (e *PackElement) GetPack() Pack {
+	return e.Pack
+}
