+++ packs/ruby/analysis.go
@@ -1,0 +1,10 @@
+package ruby
+
+import "github.com/cloud66/starter/packs"
+
+type Analysis struct {
+	packs.AnalysisBase
+
+	ServiceYAMLContext *ServiceYAMLContext
+	DockerfileContext  *DockerfileContext
+}
