+++ packs/ruby/dockerfile-context.go
@@ -1,0 +1,7 @@
+package ruby
+
+import "github.com/cloud66/starter/packs"
+
+type DockerfileContext struct {
+	packs.DockerfileContextBase
+}
