+++ packs/ruby/dockerfile-writer.go
@@ -2,10 +2,6 @@ package ruby
 
 import "github.com/cloud66/starter/packs"
 
-type DockerfileContext struct {
-	packs.DockerfileContextBase
-}
-
 type DockerfileWriter struct {
 	packs.DockerfileWriterBase
 }
