+++ packs/ruby/service-yaml-writer.go
@@ -2,14 +2,6 @@ package ruby
 
 import "github.com/cloud66/starter/packs"
 
-type ServiceYAMLContext struct {
-	packs.ServiceYAMLContextBase
-}
-
 type ServiceYAMLWriter struct {
 	packs.ServiceYAMLWriterBase
-}
-
-func (w *ServiceYAMLWriter) Write(context *ServiceYAMLContext) error {
-	return w.ServiceYAMLWriterBase.Write(context)
 }
