+++ packs/service-yaml-context-base.go
@@ -1,0 +1,8 @@
+package packs
+
+import "github.com/cloud66/starter/common"
+
+type ServiceYAMLContextBase struct {
+	Services []*common.Service
+	Dbs      []string
+}
